Add TestResultFrame parent class for test results
